requirements.txt
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
Flask-Login==0.6.3
Flask-Migrate==4.1.0
Flask-Limiter==3.12
Werkzeug==3.1.3
gunicorn==23.0.0
PyMySQL==1.1.1
pytest==8.3.5
